// src/app/models/parametrage-plan-data.ts
/**
 * Classe pour le paramétrage de plan
 */
export class ParametragePlanData
{
    planId: string;
    typePlan: string;
    description: string; 
    contrat: string;
    typeContrat: string;
    college: string;
    regime: string;
    typeCotisation: string;

    constructor() { }

    

    /*updatePlan()
    {
        console.log("Update plan : " + this.planId + " - " + this.typePlan + " - " + this.description + " - " + this.contrat + " - " + this.typeContrat + " - " + this.college + " - " + this.regime + " - " + this.typeCotisation);
    }*/
}